test: "FourBasicOperationTest"
